+++ src/net/src/net_pubsub.cc
@@ -298,7 +298,8 @@ void PubSubThread::PubSubChannels(const std::string& pattern, std::vector<std::s
   } else {
     std::lock_guard l(channel_mutex_);
     for (auto& channel : pubsub_channel_) {
-      if (pstd::stringmatchlen(channel.first.c_str(), channel.first.size(), pattern.c_str(), pattern.size(), 0)) {
+      if (pstd::stringmatchlen(channel.first.c_str(), static_cast<int32_t>(channel.first.size()), pattern.c_str(),
+                               static_cast<int32_t>(pattern.size()), 0)) {
         if (!channel.second.empty()) {
           result->push_back(channel.first);
         }
@@ -315,7 +316,7 @@ void PubSubThread::PubSubNumSub(const std::vector<std::string>& channels,
     subscribed = 0;
     for (auto& channel : pubsub_channel_) {
       if (channel.first == i) {
-        subscribed = channel.second.size();
+        subscribed = static_cast<int32_t>(channel.second.size());
       }
     }
     result->push_back(std::make_pair(i, subscribed));
@@ -326,7 +327,7 @@ int PubSubThread::PubSubNumPat() {
   int subscribed = 0;
   std::lock_guard l(pattern_mutex_);
   for (auto& channel : pubsub_pattern_) {
-    subscribed += channel.second.size();
+    subscribed += static_cast<int32_t>(channel.second.size());
   }
   return subscribed;
 }
@@ -402,8 +403,9 @@ void* PubSubThread::ThreadMain() {
 
           // Send message to a channel pattern's clients
           pattern_mutex_.lock();
-          for (auto & it : pubsub_pattern_) {
-            if (pstd::stringmatchlen(it.first.c_str(), it.first.size(), channel.c_str(), channel.size(), 0)) {
+          for (auto& it : pubsub_pattern_) {
+            if (pstd::stringmatchlen(it.first.c_str(), static_cast<int32_t>(it.first.size()), channel.c_str(),
+                                     static_cast<int32_t>(channel.size()), 0)) {
               for (size_t i = 0; i < it.second.size(); i++) {
                 if (!IsReady(it.second[i]->fd())) {
                   continue;
